% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demItaly.R
\docType{package}
\name{demItaly}
\alias{demItaly}
\alias{demItaly-package}
\title{demItaly: Italian demographic data.}
\description{
The datasets are used  by packages 'demdata', 'demest',
'demlife', 'demfam', and 'classconc'.  Most datasets are arrays,
though some are data.frames.  Most of the data is real,
but some is simulated.
}
