% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italy-births-reg.R
\docType{data}
\name{italy.births.reg}
\alias{italy.births.reg}
\title{Registered births in Italy by age and region, 2006-2016.}
\format{An array with dimensions "region", "age" and "year"}
\source{
Custom tabulation from Italian Statistical Institute (ISTAT)
\url{http://demo.istat.it/}.
}
\usage{
italy.births.reg
}
\description{
The \code{"age"} variable refers to the age of the mother.
}
\details{
All births to mothers aged less than 20 (included those to mothers aged less than
15) have been included in age group \code{"15-19"}, and all births to mothers aged
45 or higher (including those to mothers aged 50+) have been included in age
group \code{"45-49"}.
}
\keyword{datasets}
