% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italy-deaths-reg.R
\docType{data}
\name{italy.deaths.reg}
\alias{italy.deaths.reg}
\title{Registered deaths in Italy by age, sex, and region, 2006-2015.}
\format{An array with dimensions "age", "sex", "region", and "year"}
\source{
Custom tabulation from Italian Statistical Institute (ISTAT)
\url{http://dati.istat.it/}.
}
\usage{
italy.deaths.reg
}
\description{
Registered deaths in Italy by age, sex, and region, 2006-2015.
}
\seealso{
\code{\link{italy.popn.reg}}, \code{\link{italy.births.reg}},
\code{\link{nz.int.mig.reg}}, \code{\link{nz.ext.mig.reg}}.
}
\keyword{datasets}
