% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italy-ext-imm.R
\docType{data}
\name{italy.ext.imm}
\alias{italy.ext.imm}
\title{New registrations of people coming from a foreing country
per region of destination, sex, large age groups, year 2006-2014.}
\format{An array with dimensions "region", "sex", "age"  and "year"}
\source{
Custom tabulation from Italian Statistical Institute (ISTAT)
\url{http://dati.istat.it/}.
}
\usage{
italy.ext.imm
}
\description{
New registrations of people coming from a foreing country
per region of destination, sex, large age groups, year 2006-2014.
}
\keyword{datasets}
