% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italy-intl-emi.R
\docType{data}
\name{italy.intl.emi}
\alias{italy.intl.emi}
\title{Cancellations of people leaving from an italian region to any other
italian region per departing region, sex, large age groups,
year 2006-2014.}
\format{An array with dimensions "region", "sex", "age"  and "year"}
\source{
Custom tabulation from Italian Statistical Institute (ISTAT)
\url{http://dati.istat.it/}.
}
\usage{
italy.intl.emi
}
\description{
Cancellations of people leaving from an italian region to any other
italian region per departing region, sex, large age groups,
year 2006-2014.
}
\keyword{datasets}
