% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italy-intl-imm.R
\docType{data}
\name{italy.intl.imm}
\alias{italy.intl.imm}
\title{New registrations of people coming from other italian regions
per region of destination, sex, large age groups, year 2006-2014.}
\format{An array with dimensions "region", "sex", "age"  and "year"}
\source{
Custom tabulation from Italian Statistical Institute (ISTAT)
\url{http://dati.istat.it/}.
}
\usage{
italy.intl.imm
}
\description{
New registrations of people coming from other italian regions
per region of destination, sex, large age groups, year 2006-2014.
}
\keyword{datasets}
