% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italy-popn1Y.R
\docType{data}
\name{italy.popn.reg.1Y}
\alias{italy.popn.reg.1Y}
\title{Estimates of Italian population by age, sex, and region, 2006-2014.}
\format{An array with dimensions "age", "sex", "region", and "year".
Single year age groups.}
\source{
Custom tabulation from Italian Statistical Institute (ISTAT)
\url{http://demo.istat.it/}.
}
\usage{
italy.popn.reg.1Y
}
\description{
The population counts refer to 1st January of each year.
}
\details{
Census data for year 2011
}
\keyword{datasets}
