% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italy-popn5Y-100.R
\docType{data}
\name{italy.popn5Y.100}
\alias{italy.popn5Y.100}
\title{Estimates of Italian population by age, sex, and region, 2006-2014.}
\format{An array with dimensions "age", "sex", "region", and "year".
5 years age groups. Last groups "100 +".}
\source{
Custom tabulation from Italian Statistical Institute (ISTAT)
\url{http://demo.istat.it/}.
}
\usage{
italy.popn5Y.100
}
\description{
The population counts refer to 1st January of each year.
}
\details{
Census data for year 2011
}
\keyword{datasets}
